chore: configure files structure and rename comps
