+++ src/pages/social/explore.tsx
@@ -1,8 +1,8 @@
 import type { ReactElement } from "react";
 
 import type { NextPageWithLayout } from "../_app";
-import PageLayout from "@/containers/social/SocialLayout";
-import PostExploration from "@/containers/social/PostExploration";
+import PageLayout from "@/containers/social/Layout";
+import PostExploration from "@/containers/social/exploration";
 
 const ExplorePage: NextPageWithLayout = () => {
   return <PostExploration />;
